requirements.txt
 #==============================================================================
# FILE: requirements.txt
# ==============================================================================
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-generativeai==0.8.3
slack-bolt==1.18.0
aiohttp==3.9.0
aiosqlite==0.19.0
apscheduler==3.10.4
python-dotenv==1.0.0
pyyaml==6.0.1
jira==3.5.2
slack-sdk==3.27.0
